bin/setup.py: Drop dl layers from layerindex
Issue: LINUXEXEC-31709

The dl layers only provide sources for building, they don't provide meta files
such as bb, bbclass or bsp, so they don't have to be in layerindex, use repo
command to manage them as a git repo can make the process easier.

The workflow is:
* Get dl layers from data/json/dl_layer_recommends.json when --dl-layers is specified
* Add dl layers to default.xml

Then they don't have to be in layerindex.

Signed-off-by: Robert Yang <liezhi.yang@windriver.com>