setup.sh: add proxy settings to .gitconfig
Issue: LIN1024-3496

The repo uses .gitconfig in the project directory, which is generated by
setup.sh. The add_gitconfig function in setup.sh reads the global
settings and copies some settings to .gitconfig. But when the user uses
a proxy in git, the proxy settings are not copied to .gitconfig. This
will cause repo to not work properly. We need to check if there are
proxy settings in git, and copy to .gitconfig if they exist.

Signed-off-by: Yi Zhao <yi.zhao@windriver.com>
Signed-off-by: Changqing Li <changqing.li@windriver.com>